requirements.txt
python-telegram-bot==21.0.1
flask==2.3.2
requests==2.31.0
python-dotenv==1.0.0